Assignment mostly completed
Upload test to make sure I have something to submit before the deadline. I still have a bug in my code, not sure how to check if the matrix is empty or not. Still have some things to try but better get something submitted before the deadline as it is very tight.