Resubmit chachematrix after feedback
I had some negative feedback on my code with one reviewer believing that my code didn't do what it was meant to, although when I tested it it was working fine. I think the problems I had related to the code I was using to test rather than the code itself. Hopefully this works better now.